Modified to use TRIE index in AzaMaster.
